refactor(n): extract ProfileTab, BillingTab, SettingsTab, TeamTab from advisor portal
page.tsx (2,761 LOC on the branch pre-N-03 → 805 LOC now) was a single-file monolith
that mixed session management, 8 tab views, firm management, dispute handling, and
notifications.  The four remaining inline tabs were blocking code review, slowing IDE
symbol resolution, and making it harder to understand the session-vs-tab boundary.

Extracted components:
- ProfileTab.tsx (228 LOC) — profile fields + photo upload + special offer + reviews.
  Internalises savingProfile/profileSaved state; parent passes advisor + reviews +
  onAdvisorChange callback.
- BillingTab.tsx (226 LOC) — credit pack purchasing, featured-advisor upsell, lead
  billing history, top-up history.  Internalises topupHistory + useEffect fetch so the
  parent no longer needs loadTopupHistory().
- SettingsTab.tsx (119 LOC) — email notification prefs + KYC link + account info.
  Internalises notifPrefs/savingNotifs/notifSaved/loadNotifPrefs/saveNotifPrefs; parent
  only passes advisor.
- TeamTab.tsx (552 LOC) — firm member list, invite flow, firm analytics, seat-request.
  All firm-specific state (firmMembers, firmInvites, firmDetails, firmMemberCount,
  firmAnalytics, firmTab, editingFirm, seatRequestSeats, inviteEmail, etc.) now lives
  inside TeamTab; parent renders <TeamTab advisor={advisor} />.

page.tsx retains: session/auth logic, shared data (leads/stats/billing/reviews/
viewsByDay/categoryPricing), dispute modal (cross-tab, triggered from LeadsTab), and
the outer shell + nav.

Caller verification:
- ProfileTab: sole render site is view==="profile" && advisor in page.tsx.
- BillingTab: sole render site is view==="billing" in page.tsx; billing data from
  loadData() passed as prop; topupHistory fetched independently on mount.
- SettingsTab: sole render site is view==="settings" in page.tsx.  Pre-existing
  notifPrefsLoaded guard converted to useEffect once-on-mount pattern — same semantics.
- TeamTab: sole render site is view==="team" && isFirmAdmin in page.tsx.  loadFirmData
  called on mount via useEffect; loadFirmAnalytics lazy via sub-tab click — same
  trigger timing as before.

The nav click handler no longer calls loadFirmData/loadTopupHistory/loadNotifPrefs;
each tab self-loads on mount, achieving equivalent first-render semantics with simpler
orchestration.

Queue item N-03c.  Audit §6.

Refs: #242
Audit: docs/audits/codebase-health-2026-04-24.md §6 (file splits + UX)
Queue: docs/audits/REMEDIATION_QUEUE.md · N-03c
